# refactor(log): elide keyval translation to map